backend/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
jedi==0.18.2
parso==0.8.3
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pytz==2022.7
sqlparse==0.4.3
